% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtool.R
\name{c_d}
\alias{c_d}
\title{Change working directory using project root (here::here()) with glob support}
\usage{
c_d(folder = NULL, ls = FALSE, show = TRUE)
}
\arguments{
\item{folder}{Folder name, glob, relative path, or absolute path. NULL -> go to root.}

\item{ls}{If TRUE, list directory after changing.}

\item{show}{If TRUE, print message.}
}
\description{
Change working directory using project root (here::here()) with glob support
}
