% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtool.R
\name{c_dir}
\alias{c_dir}
\title{Create folder and/or file relative to project root}
\usage{
c_dir(folder, file = NULL, recursive = TRUE, overwrite = FALSE, show = TRUE)
}
\description{
Create folder and/or file relative to project root
}
