% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtool.R
\name{d_ir}
\alias{d_ir}
\title{List files/dirs with terminal-like glob support}
\usage{
d_ir(
  path = ".",
  pattern = NULL,
  full = FALSE,
  recursive = FALSE,
  all.files = FALSE,
  type = c("all", "file", "dir")
)
}
\description{
List files/dirs with terminal-like glob support
}
