% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtool.R
\name{l_s}
\alias{l_s}
\title{Quick list (alias of d_ir)}
\usage{
l_s(
  path = ".",
  pattern = NULL,
  full = FALSE,
  recursive = FALSE,
  all.files = FALSE,
  type = c("all", "file", "dir")
)
}
\description{
Quick list (alias of d_ir)
}
