% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtool.R
\name{r_name}
\alias{r_name}
\title{Rename file inside a folder relative to project root}
\usage{
r_name(folder, old, new, overwrite = FALSE, show = TRUE)
}
\description{
Rename file inside a folder relative to project root
}
