% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtool.R
\docType{package}
\name{rtool-package}
\alias{rtool}
\alias{rtool-package}
\title{rtool: Project helpers for navigation and file ops}
\description{
A small set of helpers for navigating project roots (via here::here()),
listing, searching, opening, creating, and renaming files.
}
\author{
\strong{Maintainer}: Do Thanh Liem \email{liem20k@gmail.com}

}
\keyword{internal}
